+++ webpack.config.js
@@ -30,10 +30,8 @@ const config = {
     rules: [
       {
         test: /\.js$/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader'
-        }
+        loader: 'babel-loader',
+        exclude: /node_modules/
       }
     ]
   }